- Update getOrHeadHandler to dispatch to getOrHeadHandlerUnixFs for Unixfs reponse format - Write functions for logic in getOrHeadHandler, to make it easier to read and to enable reuse in getOrHeadHandlerUnixfs - Move Unixfs specific functions to gateway_handler_unixfs.go - Check for _redirects file if we have origin isolation
