+++ test/sharness/t0114-gateway-subdomains.sh
@@ -145,6 +145,42 @@ test_expect_success 'start daemon with empty config for Gateway.PublicGateways' 
 '
 
 ## ============================================================================
+## Test _redirects file support
+## ============================================================================
+
+# Directory tree crafted to test _redirects file support
+test_expect_success "Add the _redirects file test directory" '
+  mkdir -p testredirect/ &&
+  echo "index.html" > testredirect/index.html &&
+  echo "one.html" > testredirect/one.html &&
+  echo "two.html" > testredirect/two.html &&
+  echo "^/301-redirect-one$ /one.html 301" > testredirect/_redirects &&
+  echo "^/302-redirect-two$ /two.html 302" >> testredirect/_redirects &&
+  echo "^/200-index$ /index.html 200" >> testredirect/_redirects &&
+  REDIRECTS_DIR_CID=$(ipfs add -Qr --cid-version 1 testredirect)
+'
+
+REDIRECTS_DIR_HOSTNAME="${REDIRECTS_DIR_CID}.ipfs.localhost:$GWAY_PORT"
+
+test_expect_success "_redirects - /301-redirect-one" '
+  curl -sD - --resolve $REDIRECTS_DIR_HOSTNAME:127.0.0.1 "http://$REDIRECTS_DIR_HOSTNAME/301-redirect-one" > response &&
+  test_should_contain "one.html" response &&
+  test_should_contain "301 Moved Permanently" response
+'
+
+test_expect_success "_redirects - /302-redirect-two" '
+  curl -sD - --resolve $REDIRECTS_DIR_HOSTNAME:127.0.0.1 "http://$REDIRECTS_DIR_HOSTNAME/302-redirect-two" > response &&
+  test_should_contain "two.html" response &&
+  test_should_contain "302 Found" response
+'
+
+test_expect_success "_redirects - /200-index" '
+  curl -sD - --resolve $REDIRECTS_DIR_HOSTNAME:127.0.0.1 "http://$REDIRECTS_DIR_HOSTNAME/200-index" > response &&
+  test_should_contain "index.html" response &&
+  test_should_contain "200 OK" response
+'
+
+## ============================================================================
 ## Test path-based requests to a local gateway with default config
 ## (forced redirects to http://*.localhost)
 ## ============================================================================
